requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
dateutils==0.6.12
Django==3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-oauth-toolkit==1.2.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jdatetime==4.1.0
mongoengine==0.24.1
mongoengine-goodjson==1.1.8
mysqlclient==2.1.1
oauthlib==3.2.2
PyJWT==2.6.0
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
